tests/wpt/tests/dom/events/scrolling/scroll-event-fired-to-element.html

<!doctype html>
<meta charset=utf-8>
<title>Scroll event should behave correctly for Element.offsetTop and Element.offsetLeft</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<link rel="help" href="https://drafts.csswg.org/cssom-view/#extensions-to-the-htmlelement-interface">
<link rel="help" href="https://drafts.csswg.org/cssom-view/#scrolling-events">
<div id=log></div>
<div id="container">
</div>
<script>

function setupTarget() {
  var container = document.getElementById("container");
  container.innerHTML = "";

  var target = document.createElement("div");
  var overflowing_child = document.createElement("div");

  target.style = "overflow:scroll; height: 100px; width: 100px; scrollbar-width: none";
  overflowing_child.style = "height: 200px; width: 200px;";
  target.appendChild(overflowing_child);
  container.appendChild(target);
  return target;
}

function triggerReflow(target) {
  // These should trigger reflow
  target.style.display = "inline-block";
}

// Create a promise, returning scrollTop whenever it receive a "scroll" event.
function promiseForScrollTop(target) {
  return new Promise(
    resolve =>
      target.addEventListener("scroll", () => resolve(target.scrollTop)),
    { once: true },
  );
}

// Create a promise, returning scrollLeft whenever it receive a "scroll" event.
function promiseForScrollLeft(target) {
  return new Promise(
    resolve =>
      target.addEventListener("scroll", () => resolve(target.scrollLeft)),
    { once: true },
  );
}

promise_test(async (t) => {
  var target = setupTarget();

  assert_equals(target.scrollTop, 0);
  var updatedOffsetTop = promiseForScrollTop(target);
  target.scrollTop = 10;
  assert_equals(await updatedOffsetTop, 10);

  assert_equals(target.scrollLeft, 0);
  var updatedOffsetLeft = promiseForScrollLeft(target);
  target.scrollLeft = 10;
  assert_equals(await updatedOffsetLeft, 10);

}, "scrollTop and scrollLeft should fire scroll event.");

promise_test(async (t) => {
  var target = setupTarget();

  target.addEventListener("scroll", () => assert_unreached("Any scroll event should not be observed"));

  assert_equals(target.scrollTop, 0);
  target.scrollTop = 0;

  assert_equals(target.scrollLeft, 0);
  target.scrollLeft = 0;

  triggerReflow(target); // Ensure that all event is flushed.
}, "scrollTop and scrollLeft being set with the same value.");

promise_test(async (t) => {
  var target = setupTarget();

  target.addEventListener("scroll", () => assert_unreached("Any scroll event should not be observed"));

  target.scrollTop = -100;
  target.scrollLeft = -100;

  triggerReflow(target); // Ensure that all event is flushed.
}, "scrollTop and scrollLeft being set with invalid scroll offset.");

promise_test(async (t) => {
  var target = setupTarget();

  assert_equals(target.scrollTop, 0);
  var updatedOffsetTop = promiseForScrollTop(target);
  target.scrollTop = 1000;
  assert_equals(await updatedOffsetTop, 100);

  assert_equals(target.scrollLeft, 0);
  var updatedOffsetLeft = promiseForScrollLeft(target);
  target.scrollLeft = 1000;
  assert_equals(await updatedOffsetLeft, 100);

  target.addEventListener("scroll", () => assert_unreached("Any scroll event should not be observed"));

  target.scrollTop = 1000;
  target.scrollLeft = 1000;

  triggerReflow(target); // Ensure that all event is flushed.
}, "scrollTop and scrollLeft when scrolling above maximum offset.");

</script>
